atualizando codebase

+++ src/main/java/br/com/marcospcruz/gestorloja/controller/FabricanteController.java
@@ -1,6 +1,8 @@
 package br.com.marcospcruz.gestorloja.controller;
 
+import java.util.ArrayList;
 import java.util.List;
+import java.util.Map;
 
 import br.com.marcospcruz.gestorloja.dao.Crud;
 import br.com.marcospcruz.gestorloja.dao.CrudDao;
@@ -9,26 +11,29 @@ import br.com.marcospcruz.gestorloja.model.Operacao;
 
 public class FabricanteController extends ControllerBase {
 	private static final String FABRICANTE_INVALIDO = "Fabricante inválido!";
-	private Crud<Fabricante> fabricanteDao;
+	// private Crud<Fabricante> fabricanteDao;
 	private List<Fabricante> fabricantes;
 	private Fabricante fabricante;
 
 	public FabricanteController() {
 		super();
-		fabricanteDao = new CrudDao<>();
+		// fabricanteDao = new CrudDao<>();
 	}
 
 	@Override
 	public void busca(Object id) throws Exception {
-		fabricanteDao = new CrudDao<>();
+
+		Crud<Fabricante> fabricanteDao = new CrudDao<>();
 		fabricante = fabricanteDao.busca(Fabricante.class, new Integer(id.toString()));
 
 	}
 
 	@Override
 	public List buscaTodos() {
-		if (fabricantes == null || fabricantes.isEmpty())
+		if (fabricantes == null || fabricantes.isEmpty()) {
+			Crud<Fabricante> fabricanteDao = new CrudDao<>();
 			fabricantes = fabricanteDao.busca("fabricante.buscaTodos");
+		}
 		return getList();
 	}
 
@@ -40,9 +45,11 @@ public class FabricanteController extends ControllerBase {
 
 	@Override
 	public void busca(String text) throws Exception {
-		String nomeFabricante = "%"+text.toUpperCase()+"%";
-		fabricante = fabricanteDao.busca("fabricante.readParametroLike", "nome", nomeFabricante);
-
+		// String nomeFabricante = "%"+text.toUpperCase()+"%";
+		// fabricante = fabricanteDao.busca("fabricante.readParametroLike", "nome",
+		// nomeFabricante);
+		int id = Integer.parseInt(text);
+		busca(id);
 	}
 
 	@Override
@@ -65,7 +72,21 @@ public class FabricanteController extends ControllerBase {
 
 	@Override
 	public void excluir() throws Exception {
-		fabricanteDao.delete(fabricante);
+		if (!fabricante.getItensEstoque().isEmpty())
+			throw new Exception("Exclusão inválida. Há " + fabricante.getItensEstoque().size()
+					+ " ítens deste Fabricante / Marca no Estoque.");
+
+		// busca(fabricante.getIdFabricante());
+		Crud<Fabricante> fabricanteDao = new CrudDao<>();
+		try {
+			fabricante = fabricanteDao.busca(Fabricante.class, fabricante.getIdFabricante());
+			fabricanteDao.delete(fabricante);
+			Map<Object, Object> cache = getCacheMap();
+			cache.remove(fabricante.getIdFabricante());
+		} catch (IllegalArgumentException e) {
+
+			throw new Exception("Falha ao excluir dados.");
+		}
 
 	}
 
@@ -83,8 +104,10 @@ public class FabricanteController extends ControllerBase {
 
 	@Override
 	public void salva() throws Exception {
+		Crud<Fabricante> fabricanteDao = new CrudDao<>();
 		fabricante = fabricanteDao.update(fabricante);
 
+		fabricantes = new ArrayList<>();
 	}
 
 	@Override
@@ -97,12 +120,24 @@ public class FabricanteController extends ControllerBase {
 	public void validaExistente(String text) throws Exception {
 		Fabricante novo = null;
 		try {
+			Crud<Fabricante> fabricanteDao = new CrudDao<>();
 			novo = fabricanteDao.busca("fabricante.readParametroLike", "nome", "%" + text.toUpperCase() + "%");
 		} catch (Exception e) {
 
 		}
 		if (fabricante.getIdFabricante() == null && novo != null)
 			throw new Exception("Marca / Fabricante já cadastrado.");
+
+	}
+
+	@Override
+	public void novo() {
+		fabricante = new Fabricante();
+	}
+
+	public void buscaNome(String nome) {
+		Crud<Fabricante> fabricanteDao = new CrudDao<>();
+		fabricante = fabricanteDao.busca("fabricante.readNome", "nome", nome.toUpperCase());
 
 	}
